Ability to render columns using the entire object
This extension allows users to create links in a column using data that isn't present in the immediate column. The user can have a user.alias column that links to the profile page using the user.id, e.g. `renderWithData: (user) => { return <a href={'users/' + user.id}>{user.alias}</a> }`